Migrate coverage check in its own file and improve it.
It's already large and it will grow larger.
Add filtering on ignored files. But it's still using the wrong total, this will
be fixed later.
Improve sorting to be base on source file name, then symbol name and only then
after on line number.
Rename many symbols to be clearer.