Get rid of "go tool cover" dependency.
Use the local code instead, this permits reducing the amount of disk I/O since
the in-memory cache of the files is used instead of re-reading the files from
disk and reduces the number of processes started during coverage. Processes are
slow on Windows and creates overhead, even on linux.

Fix bug in tests were change.Package() wouldn't be set because $GOPATH was
replaced too late.