Finish the migration of definitions back into checks.
Background: I initially thought it was cool to use godoc to document the
configuration file. This permits to document once and the documentation is the
code. In practice, this didn't work out that well because it created very busy
godoc file and it was harder to relate to the rendered yaml. So I wrote manually
CONFIGURATION.md instead, which should serve users much better.